requirements.txt
Flask
pyzbar
Pillow
Werkzeug
opencv-python  # 如果需要使用OpenCV
